Refactor settings panel logic into focused hooks
